examine: Examines power, Type S, and Type M error for a given design

Calculates Power, Type S, and Type M error and returns them in a list or df, depending on whether a single true effect size or range is provided. retro_design() is faster as it uses the closed form solution from Lu et al. (2018), but this function can be used for t distributions, whereas retro_design() cannot. Function originally provided in Gelman and Carlin (2014), reused with permission.

Usage

examine(
  effect_size,
  standard_error,
  alpha = 0.05,
  df = Inf,
  number_simulations = 10
)

Arguments

effect_size

a numeric or list containing an estimate(s) of the "true" effect size.

standard_error

a numeric, standard error of the estimate

alpha

a numeric, the statistical significance threshold

df

a numeric, the degrees of freedom. df=Inf is equivalent to a normal distribution.

number_simulations

a numeric, how many times to simulate when calculating Type M error.

Value

either a list of length 3 containing the power, type s, and type M error, or if A is a list, a df that is 4 by length(A), with an effect size and it's corresponding power, type s, and type m errors in each row.

Examples

examine(1,3.28)
#> $power
#> [1] 0.06071472
#> 
#> $typeS
#> [1] 0.1937123
#> 
#> $exaggeration
#> [1] 7.965978
#> 
examine(list(.2,2,20),8.1)
#> Error in UseMethod("examine", effect_size): no applicable method for 'examine' applied to an object of class "list"
examine(.5,1,df=10)
#> $power
#> [1] 0.06795991
#> 
#> $typeS
#> [1] 0.1564247
#> 
#> $exaggeration
#> [1] NaN
#>